+++ src/Cloud/DTO/DatabaseDependencyType.php
@@ -8,6 +8,7 @@ enum DatabaseDependencyType: string
 {
     case ENVIRONMENT_ATTACHMENT = 'environment_attachment';
     case OWNED_DATABASE_CHILD = 'owned_database_child';
+    case DERIVED_PARENT_DEPENDENCY = 'derived_parent_dependency';
     case UNMANAGED_DATABASE_CHILD = 'unmanaged_database_child';
     case OWNERSHIP_CONFLICT = 'ownership_conflict';
     case DATABASE_SNAPSHOT = 'database_snapshot';